added rest apis for updating imports

+++ src/main/java/com/iexceed/importupdater/fileutil/UpdateALLImports.java
@@ -11,25 +11,28 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
 import com.iexceed.importupdater.model.ImportDetails;
 import com.iexceed.importupdater.model.ImportsData;
 import com.iexceed.importupdater.seeddata.SeedData;
+import com.iexceed.importupdater.shutdown.AppShutDownUtil;
 
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
+@Service
 public class UpdateALLImports {
 
-	public static void init(String basePath) {
+	static String[] pathsToUpdate = { "Sources/Containers/AndroidStudio/app/src/main/java/com",
+			"Sources/Plugins/AndroidStudio/app/src/main/java/com" };
+
+	public static void initCMD(String basePath) {
 		Assert.notNull(basePath, "Base path cant be null");
-		
+
 		SeedData seedData = new SeedData();
 		ImportsData importsData = seedData.loadImportsData();
-
-		String[] paths = { "Sources/Containers/AndroidStudio/app/src/main/java/com",
-				"Sources/Plugins/AndroidStudio/app/src/main/java/com" };
 
 		if (!basePath.endsWith("/")) {
 			basePath += "/";
@@ -42,10 +45,38 @@ public class UpdateALLImports {
 			importsMap.put(imprt.searchFor, imprt.replace);
 		}
 
-		for (String path : paths) {
-			updateImports(basePath + path, importsMap);
+		start(basePath, pathsToUpdate, importsMap);
+
+	}
+
+	public static void initREST(String basePath, String[] pathsArr, Map<String, String> importsMap) {
+
+		if (!basePath.endsWith("/")) {
+			basePath += "/";
 		}
 
+		if (pathsArr != null) {
+			pathsToUpdate = pathsArr;
+		}
+
+		start(basePath, pathsToUpdate, importsMap);
+
+	}
+
+	private static void start(String basePath, String[] pathsArr, Map<String, String> importsMap) {
+		String fullPath = "";
+		log.info("Init done..");
+		for (String path : pathsArr) {
+
+			if (path.startsWith("/")) {
+				fullPath = basePath + (path.length() > 1 ? path.substring(1) : path); // ex: path = "/"
+			} else {
+				fullPath = basePath + path;
+			}
+
+			updateImports(fullPath, importsMap);
+		}
+		AppShutDownUtil.shutDown(0); // Stop
 	}
 
 	private static void updateImports(String path, Map<String, String> importsMap) {
@@ -69,6 +100,8 @@ public class UpdateALLImports {
 
 			String line = null;
 
+			boolean isFileChanged = false;
+
 			for (int i = 0; i < lines.size(); i++) {
 				line = lines.get(i).trim();
 
@@ -76,21 +109,25 @@ public class UpdateALLImports {
 					if (importsMap.containsKey(line)) {
 						line = line.replaceAll(line, importsMap.get(line));
 						lines.set(i, line);
+						isFileChanged = true;
 					}
 				} else {
 					if (importsMap.containsKey(line)) {
 						line = line.replaceAll(line, importsMap.get(line));
 						lines.set(i, line);
+						isFileChanged = true;
 					}
 				}
 
 			}
 
-			Path status = Files.write(filePath, lines);
-			if (status != null) {
-				log.info("File updated successfully");
-			} else {
-				log.info("File failed to update at location {} ", filePath.toAbsolutePath().toString());
+			if (isFileChanged) {
+				Path status = Files.write(filePath, lines);
+				if (status != null) {
+					log.info("File updated successfully at location {} ", filePath.toAbsolutePath().toString());
+				} else {
+					log.info("File failed to update at location {} ", filePath.toAbsolutePath().toString());
+				}
 			}
 		} catch (IOException e) {
 			e.printStackTrace();
@@ -106,7 +143,7 @@ public class UpdateALLImports {
 		try (Stream<Path> walk = Files.walk(path)) {
 
 			files = walk.filter(Files::isRegularFile).map(x -> x.toString()).collect(Collectors.toList());
-			log.info("Found {} files", files.size());
+			log.info("Found {} files at base path {} ", files.size(), basePath);
 
 		} catch (IOException e) {
 			e.printStackTrace();
